backend/requirements.txt
anyio==4.12.0
APScheduler==3.11.2
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
fastapi==0.124.4
greenlet==3.3.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0
python-dotenv==1.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
tornado==6.5.4
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.38.0
alembic==1.17.2
tzdata==2024.1